Move ipc types into separate module
